prompt: fix "%" before prompt on Windows Console
On the Windows Console, the software ConEmu and others are based on,
a reverse-video percent ('%') character is output before each prompt.
This is an indication that somehow an unterminated line is being sent to
the terminal at the end of each output cycle. This small change fixes
this without fussing with the configuration on other platforms.